Implements PR feedback less deleteMany
