+++ src/packages/admin-ui-components/README.md
@@ -23,19 +23,17 @@ const MyPage = () => {
 };
 ```
 
-## <Dropdown />
+## <Popover />
 
-A Dropdown Menu
+A Popover Menu
 
 ```tsx
-import { DropdownItem, Dropdown } from '@exogee/graphweaver-admin-ui-components';
+import { PopoverItem, Popover } from '@exogee/graphweaver-admin-ui-components';
 
 const MyPage = () => {
-	const externalLinkItems: DropdownItem[] = [];
-	
-	return (
-		<Dropdown items={externalLinkItems}>Links</Dropdown>
-	);
+	const externalLinkItems: PopoverItem[] = [];
+
+	return <Popover items={externalLinkItems}>Links</Popover>;
 };
 ```
 
@@ -44,4 +42,5 @@ const MyPage = () => {
 Comprehensive documentation and usage examples can be found on our [Docs Site](https://graphweaver.com/docs). It covers installation instructions, detailed API documentation, and guides to help you get started with Graphweaver.
 
 ## Graphweaver CLI `graphweaver`
+
 The Graphweaver Command Line Interface (CLI) tool enables you to set up and manage Graphweaver using commands in your command-line shell. Check the `graphweaver` npm package [here.](https://www.npmjs.com/package/graphweaver)
