--- src/packages/admin-ui-components/src/dropdown/index.ts
@@ -1,2 +1,0 @@
-export type { DropdownItem, DropdownProps } from './dropdown';
-export { Dropdown } from './dropdown';
