+++ src/packages/admin-ui-components/src/popover/index.ts
@@ -1,0 +1,2 @@
+export type { PopoverItem, PopoverProps } from './component';
+export { Popover } from './component';
