+++ src/packages/admin-ui-components/src/popover/styles.module.css
@@ -1,4 +1,4 @@
-.dropdownList {
+.popoverList {
 	position: absolute;
 	display: flex;
 	flex-direction: column;
@@ -18,7 +18,7 @@
 	/* min-width: 140px; */
 }
 
-.dropdown li {
+.popover li {
 	padding: 8px 16px 8px 16px;
 	height: 32px;
 	display: block;
@@ -32,15 +32,15 @@
 	font-weight: 500;
 }
 
-.dropdown li a {
+.popover li a {
 	font-weight: 500;
 }
 
-.dropdown li p {
+.popover li p {
 	font-weight: 500;
 }
 
-.dropdown li:hover {
+.popover li:hover {
 	background: var(--primary-color);
 }
